Add files via upload
"""
Function decode can read an encoded message from a message file. File contains words associated with numbers.
To find an encoded message order numberd from smallet to biggest and arrange them into a pyramid.
Each line of the pyramid includes one more number than the line before it. Example
1
2 3
4 5 6
The numbers at the end of each line (1, 3 and 6) correspond to the words that are part of the message.
You should ignore all the other words.
"""